use id for tables and projects

+++ src/main/java/migrator/lib/uid/SessionIncrementalGenerator.java
@@ -2,20 +2,20 @@ package migrator.lib.uid;
 
 public class SessionIncrementalGenerator implements Generator {
     protected String sessionId;
-    protected Integer nextId;
+    protected Long nextId;
 
     public SessionIncrementalGenerator(String sessionId) {
-        this(sessionId, 1);
+        this(sessionId, Long.valueOf(1));
     }
 
-    public SessionIncrementalGenerator(String sessionId, Integer nextId) {
+    public SessionIncrementalGenerator(String sessionId, Long nextId) {
         this.sessionId = sessionId;
         this.nextId = nextId;
     }
     
     @Override
     public String next() {
-        Integer currentId = this.nextId;
+        Long currentId = this.nextId;
         this.nextId++;
         return this.sessionId + "-" + currentId.toString();
     }
